requirements.txt
alabaster==0.7.10
argon2-cffi==16.3.0
arrow==0.10.0
awesome-slugify==1.6.5
Babel==2.4.0
binaryornot==0.4.3
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
colorama==0.3.9
cookiecutter==1.5.1
coverage==4.4.1
decorator==4.1.2
defusedxml==0.5.0
Django==1.10.7
django-allauth==0.32.0
django-braces==1.11.0
django-ckeditor==5.3.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.3
django-extensions==1.8.1
django-haystack==2.6.1
django-js-asset==0.1.1
django-model-utils==3.0.0
django-mptt==0.8.7
django-notifications-hq==1.3
django-redis==4.8.0
django-taggit==0.22.1
django-test-plus==1.0.18
docutils==0.13.1
factory-boy==2.8.1
Faker==0.7.18
future==0.16.0
idna==2.5
imagesize==0.7.1
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
jieba==0.38
Jinja2==2.9.6
jinja2-time==0.2.0
jsonfield==2.0.2
Markdown==2.6.8
MarkupSafe==1.0
oauthlib==2.0.2
olefile==0.44
pickleshare==0.7.4
Pillow==4.2.1
poyo==0.4.1
prompt-toolkit==1.0.15
psycopg2==2.7.3
py==1.4.34
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.2
pysolr==3.6.0
pytest==3.1.3
pytest-django==3.1.2
pytest-sugar==0.8.0
python-dateutil==2.6.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
regex==2017.7.26
requests==2.18.2
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
social-auth-core==1.4.0
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.3
termcolor==1.1.0
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
whichcraft==0.4.1
whitenoise==3.3.0
Whoosh==2.7.4
